Correct Raspbian installation description
The usage of the old URL is still permit, but for how long?

The old URL will make you install the Docker CE Edge package. Use the Debian
URL for stable.

On Raspbian (as for any other Debian derivatives), the Docker package is **not**
`docker` (this is a KDE package). You need to install `docker.io` if you want the
(albeit older) Debian/Raspbian version of Docker, `docker-engine` if you follow
Raspbian specific instructions and are fine with Edge, or `docker-ce` if you
use the Debian for ARMHF URL (it works also for Raspbian and you can chose
between Stable and Edge).

The docker-engine package contains the file `/etc/bash_completion.d/docker`
so Bash completion should be installed correctly.